chore: processing decl and modules
